improve debug messages

+++ src/plugins/sched/backfill/usage_tracker.c
@@ -201,8 +201,8 @@ ut_int_destroy(utracker_int_t ut) {
 static void
 ut_int_dump_item(ut_int_item_t *item) {
   char buff[32];
-  log("%24.24s (%zu): %d", ctime_r(&(item->start), buff),
-      item->start, item->value);
+  log("%24.24s (%lld): %d", ctime_r(&(item->start), buff),
+      (long long)item->start, item->value);
 }
